requirements.txt
beautifulsoup4==4.5.1
click==6.6
cookies==2.2.1
Flask==0.11.1
geopy==1.11.0
googlefinance==0.7
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauth2==1.9.0.post1
ordereddict==1.1
python-dotenv==0.5.1
python-forecastio==1.3.4
requests==2.11.1
responses==0.5.1
six==1.10.0
slacker==0.9.24
slacksocket==0.7
websocket-client==0.37.0
Werkzeug==0.11.10
yelp==1.0.2
